python/requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-braces==1.13.0
django-configurations==2.0
django-cors-headers==2.3.0
django-extensions==2.0.7
django-filter==1.1.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-uuidfield==0.5.0
djangorestframework==3.8.2
fcm-django==0.2.19
future==0.16.0
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
gunicorn==19.8.1
httplib2==0.11.3
idna==2.7
oauth2client==4.1.2
oauthlib==2.1.0
pika==0.12.0
Pillow==5.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
pyfcm==1.4.5
PyJWT==1.6.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rest-social-auth==1.3.1
rsa==3.4.2
serializers==0.2.4
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.23
